Move the `replace_todo` stuff out of `plugin` into its own file.
